suite

+++ domain/business-object/src/main/java/com/aliateck/fact/domaine/business/object/Prestation.java
@@ -17,5 +17,5 @@ public class Prestation {
   Client client;
   Facture facture;
   Integer tarifHT;
-  Integer delaiPaiement;
+  Long delaiPaiement;
 }
